+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## HEAD
+## 1.2.0 (2020-11-15)
 
 * Attempt support for [activerecord-sqlserver-adapter](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter) (@janko)